Fix float vs int warnings. Added keys F1 and F2 to attempt to emulate iPhone and iPad display size on the current monitor (requires a monitor which reports this information).

+++ FireWorks.h
@@ -20,7 +20,7 @@ enum EFireWorkState
 
 struct SFireWork
 {
-	int X, Y;
+	float X, Y;
 
 	EFireWorkState State;
